add SessionInfo
This is cleaner, since we only pass one type of 'sessionInfo' around in the code
(instead of a different one for the `Zermelo` constructor and the
`loginBySessionInfo` function). (yay consistency)

It also allows users to get the sessionInfo of an zermelo instance.
Instead of having to save it after calling `createSesion`.